auth.py - containes added to __init__.py
